# jumpdest: faster 