--- a/core/vm/analysis.go
+++ b/core/vm/analysis.go
@@ -24,24 +24,44 @@ func codeBitmap(code []byte) []uint64 {
 	// The bitmap is 4 bytes longer than necessary, in case the code
 	// ends with a PUSH32, the algorithm will push zeroes onto the
 	// bitvector outside the bounds of the actual code.
-	bits := make([]uint64, (len(code)+32+63)/64)
-
+	bits := make(bitvec, (len(code)+32+63)/64)
 	for pc := uint64(0); pc < uint64(len(code)); {
 		op := OpCode(code[pc])
 		pc++
 		if int8(op) < int8(PUSH1) { // If not PUSH (the int8(op) > int(PUSH32) is always false).
 			continue
 		}
-		numbits := uint64(op - PUSH1)
-		x := uint64(1) << numbits
-		x = x | (x - 1) // Smear the bit to the right
-		idx := pc / 64
-		shift := pc & 63
-		bits[idx] |= x << shift
-		if shift+shift > 64 {
-			bits[idx+1] |= x >> (64 - shift)
+		if op == PUSH1 {
+			bits.set1(pc)
+			pc += 1
+			continue
 		}
-		pc += numbits + 1
+
+		numbits := uint64(op - PUSH1 + 1)
+		bits.setN(uint64(1)<<numbits-1, pc)
+		pc += numbits
 	}
 	return bits
+}
+
+// bitvec is a bit vector which maps bytes in a program.
+// An unset bit means the byte is an opcode, a set bit means
+// it's data (i.e. argument of PUSHxx).
+type bitvec []uint64
+
+func (bits bitvec) set1(pos uint64) {
+	bits[pos/64] |= 1 << (pos % 64)
+}
+
+func (bits bitvec) setN(flag uint64, pc uint64) {
+	shift := pc % 64
+	bits[pc/64] |= flag << shift
+	if shift > 32 {
+		bits[pc/64+1] = flag >> (64 - shift)
+	}
+}
+
+// codeSegment checks if the position is in a code segment.
+func (bits bitvec) codeSegment(pos uint64) bool {
+	return bits[pos/64]&(uint64(1)<<(pos%64)) == 0
 }