--- a/core/vm/contract.go
+++ b/core/vm/contract.go
@@ -59,7 +59,7 @@ type Contract struct {
 	caller        ContractRef
 	self          libcommon.Address
 	jumpdests     *JumpDestCache // Aggregated result of JUMPDEST analysis.
-	analysis      []uint64       // Locally cached result of JUMPDEST analysis
+	analysis      bitvec         // Locally cached result of JUMPDEST analysis
 	skipAnalysis  bool
 
 	Code     []byte
@@ -72,7 +72,7 @@ type Contract struct {
 }
 
 type JumpDestCache struct {
-	*simplelru.LRU[libcommon.Hash, []uint64]
+	*simplelru.LRU[libcommon.Hash, bitvec]
 	hit, total int
 	trace      bool
 }
@@ -83,7 +83,7 @@ var (
 )
 
 func NewJumpDestCache(trace bool) *JumpDestCache {
-	c, err := simplelru.NewLRU[libcommon.Hash, []uint64](jumpDestCacheLimit, nil)
+	c, err := simplelru.NewLRU[libcommon.Hash, bitvec](jumpDestCacheLimit, nil)
 	if err != nil {
 		panic(err)
 	}
@@ -153,7 +153,7 @@ func (c *Contract) isCode(udest uint64) bool {
 		}
 		// Also stash it in current contract for faster access
 		c.analysis = analysis
-		return isCodeFromAnalysis(analysis, udest)
+		return c.analysis.codeSegment(udest)
 	}
 
 	// We don't have the code hash, most likely a piece of initcode not already
@@ -164,7 +164,7 @@ func (c *Contract) isCode(udest uint64) bool {
 		c.analysis = codeBitmap(c.Code)
 	}
 
-	return isCodeFromAnalysis(c.analysis, udest)
+	return c.analysis.codeSegment(udest)
 }
 
 // AsDelegate sets the contract to be a delegate call and returns the current