… core/vm/analysis_test.go