```
erigon
BenchmarkJumpdestOpAnalysis/1
BenchmarkJumpdestOpAnalysis/PUSH1
BenchmarkJumpdestOpAnalysis/PUSH1-16         	    1584	    737480 ns/op	1666.21 MB/s
BenchmarkJumpdestOpAnalysis/PUSH2-16         	    1430	    865776 ns/op	1419.30 MB/s
BenchmarkJumpdestOpAnalysis/PUSH3
BenchmarkJumpdestOpAnalysis/PUSH3-16         	    1856	    681953 ns/op	1801.88 MB/s
BenchmarkJumpdestOpAnalysis/PUSH4
BenchmarkJumpdestOpAnalysis/PUSH4-16         	    2613	    508173 ns/op	2418.07 MB/s
BenchmarkJumpdestOpAnalysis/PUSH5
BenchmarkJumpdestOpAnalysis/PUSH5-16         	    2718	    490443 ns/op	2505.49 MB/s
BenchmarkJumpdestOpAnalysis/PUSH6
BenchmarkJumpdestOpAnalysis/PUSH6-16         	    3424	    360401 ns/op	3409.54 MB/s
BenchmarkJumpdestOpAnalysis/PUSH7
BenchmarkJumpdestOpAnalysis/PUSH7-16         	    3712	    350119 ns/op	3509.66 MB/s
BenchmarkJumpdestOpAnalysis/PUSH8
BenchmarkJumpdestOpAnalysis/PUSH8-16         	    2949	    424004 ns/op	2898.09 MB/s
BenchmarkJumpdestOpAnalysis/PUSH9
BenchmarkJumpdestOpAnalysis/PUSH9-16         	    2954	    374748 ns/op	3279.01 MB/s
BenchmarkJumpdestOpAnalysis/PUSH10
BenchmarkJumpdestOpAnalysis/PUSH10-16        	    3481	    350005 ns/op	3510.81 MB/s
BenchmarkJumpdestOpAnalysis/PUSH11
BenchmarkJumpdestOpAnalysis/PUSH11-16        	    3384	    355399 ns/op	3457.52 MB/s
BenchmarkJumpdestOpAnalysis/PUSH12
BenchmarkJumpdestOpAnalysis/PUSH12-16        	    4138	    283746 ns/op	4330.63 MB/s
BenchmarkJumpdestOpAnalysis/PUSH13
BenchmarkJumpdestOpAnalysis/PUSH13-16        	    4386	    264606 ns/op	4643.89 MB/s
BenchmarkJumpdestOpAnalysis/PUSH14
BenchmarkJumpdestOpAnalysis/PUSH14-16        	    4820	    249625 ns/op	4922.58 MB/s
BenchmarkJumpdestOpAnalysis/PUSH15
BenchmarkJumpdestOpAnalysis/PUSH15-16        	    5197	    233415 ns/op	5264.45 MB/s
BenchmarkJumpdestOpAnalysis/PUSH16
BenchmarkJumpdestOpAnalysis/PUSH16-16        	    5764	    217792 ns/op	5642.09 MB/s
BenchmarkJumpdestOpAnalysis/PUSH17
BenchmarkJumpdestOpAnalysis/PUSH17-16        	    5629	    205038 ns/op	5993.03 MB/s
BenchmarkJumpdestOpAnalysis/PUSH18
BenchmarkJumpdestOpAnalysis/PUSH18-16        	    5427	    198611 ns/op	6186.97 MB/s
BenchmarkJumpdestOpAnalysis/PUSH19
BenchmarkJumpdestOpAnalysis/PUSH19-16        	    6058	    192175 ns/op	6394.16 MB/s
BenchmarkJumpdestOpAnalysis/PUSH20
BenchmarkJumpdestOpAnalysis/PUSH20-16        	    6588	    175843 ns/op	6988.07 MB/s
BenchmarkJumpdestOpAnalysis/PUSH21
BenchmarkJumpdestOpAnalysis/PUSH21-16        	    6864	    181912 ns/op	6754.92 MB/s
BenchmarkJumpdestOpAnalysis/PUSH22
BenchmarkJumpdestOpAnalysis/PUSH22-16        	    7698	    162865 ns/op	7544.90 MB/s
BenchmarkJumpdestOpAnalysis/PUSH23
BenchmarkJumpdestOpAnalysis/PUSH23-16        	    7945	    155962 ns/op	7878.83 MB/s
BenchmarkJumpdestOpAnalysis/PUSH24
BenchmarkJumpdestOpAnalysis/PUSH24-16        	    8514	    152739 ns/op	8045.09 MB/s
BenchmarkJumpdestOpAnalysis/PUSH25
BenchmarkJumpdestOpAnalysis/PUSH25-16        	    8088	    140037 ns/op	8774.81 MB/s
BenchmarkJumpdestOpAnalysis/PUSH26
BenchmarkJumpdestOpAnalysis/PUSH26-16        	    8492	    134629 ns/op	9127.33 MB/s
BenchmarkJumpdestOpAnalysis/PUSH27
BenchmarkJumpdestOpAnalysis/PUSH27-16        	    7946	    145155 ns/op	8465.42 MB/s
BenchmarkJumpdestOpAnalysis/PUSH28
BenchmarkJumpdestOpAnalysis/PUSH28-16        	    9704	    135830 ns/op	9046.57 MB/s
BenchmarkJumpdestOpAnalysis/PUSH29
BenchmarkJumpdestOpAnalysis/PUSH29-16        	    9570	    128563 ns/op	9557.93 MB/s
BenchmarkJumpdestOpAnalysis/PUSH30
BenchmarkJumpdestOpAnalysis/PUSH30-16        	    9352	    127212 ns/op	9659.46 MB/s
BenchmarkJumpdestOpAnalysis/PUSH31
BenchmarkJumpdestOpAnalysis/PUSH31-16        	   10000	    121840 ns/op	10085.35 MB/s
BenchmarkJumpdestOpAnalysis/PUSH32
BenchmarkJumpdestOpAnalysis/PUSH32-16        	   10000	    120040 ns/op	10236.60 MB/s
BenchmarkJumpdestOpAnalysis/JUMPDEST
BenchmarkJumpdestOpAnalysis/JUMPDEST-16      	    2410	    465018 ns/op	2642.48 MB/s
BenchmarkJumpdestOpAnalysis/STOP
BenchmarkJumpdestOpAnalysis/STOP-16          	    2298	    471917 ns/op	2603.85 MB/s
```

```
geth
BenchmarkJumpdestOpAnalysis/2
BenchmarkJumpdestOpAnalysis/PUSH1#01
BenchmarkJumpdestOpAnalysis/PUSH1#01-16      	     974	   1064055 ns/op	1154.83 MB/s
BenchmarkJumpdestOpAnalysis/PUSH2#01
BenchmarkJumpdestOpAnalysis/PUSH2#01-16      	    1363	    826018 ns/op	1487.62 MB/s
BenchmarkJumpdestOpAnalysis/PUSH3#01
BenchmarkJumpdestOpAnalysis/PUSH3#01-16      	    1663	    704174 ns/op	1745.02 MB/s
BenchmarkJumpdestOpAnalysis/PUSH4#01
BenchmarkJumpdestOpAnalysis/PUSH4#01-16      	    2330	    517340 ns/op	2375.23 MB/s
BenchmarkJumpdestOpAnalysis/PUSH5#01
BenchmarkJumpdestOpAnalysis/PUSH5#01-16      	    2583	    443455 ns/op	2770.97 MB/s
BenchmarkJumpdestOpAnalysis/PUSH6#01
BenchmarkJumpdestOpAnalysis/PUSH6#01-16      	    3338	    380416 ns/op	3230.15 MB/s
BenchmarkJumpdestOpAnalysis/PUSH7#01
BenchmarkJumpdestOpAnalysis/PUSH7#01-16      	    3772	    313992 ns/op	3913.48 MB/s
BenchmarkJumpdestOpAnalysis/PUSH8#01
BenchmarkJumpdestOpAnalysis/PUSH8#01-16      	    3208	    373352 ns/op	3291.27 MB/s
BenchmarkJumpdestOpAnalysis/PUSH9#01
BenchmarkJumpdestOpAnalysis/PUSH9#01-16      	    3652	    380888 ns/op	3226.14 MB/s
BenchmarkJumpdestOpAnalysis/PUSH10#01
BenchmarkJumpdestOpAnalysis/PUSH10#01-16     	    3212	    382528 ns/op	3212.32 MB/s
BenchmarkJumpdestOpAnalysis/PUSH11#01
BenchmarkJumpdestOpAnalysis/PUSH11#01-16     	    3326	    368253 ns/op	3336.83 MB/s
BenchmarkJumpdestOpAnalysis/PUSH12#01
BenchmarkJumpdestOpAnalysis/PUSH12#01-16     	    3267	    328764 ns/op	3737.64 MB/s
BenchmarkJumpdestOpAnalysis/PUSH13#01
BenchmarkJumpdestOpAnalysis/PUSH13#01-16     	    3217	    404658 ns/op	3036.64 MB/s
BenchmarkJumpdestOpAnalysis/PUSH14#01
BenchmarkJumpdestOpAnalysis/PUSH14#01-16     	    3697	    324470 ns/op	3787.09 MB/s
BenchmarkJumpdestOpAnalysis/PUSH15#01
BenchmarkJumpdestOpAnalysis/PUSH15#01-16     	    4089	    316770 ns/op	3879.15 MB/s
BenchmarkJumpdestOpAnalysis/PUSH16#01
BenchmarkJumpdestOpAnalysis/PUSH16#01-16     	    4597	    256662 ns/op	4787.63 MB/s
BenchmarkJumpdestOpAnalysis/PUSH17#01
BenchmarkJumpdestOpAnalysis/PUSH17#01-16     	    5588	    216130 ns/op	5685.46 MB/s
BenchmarkJumpdestOpAnalysis/PUSH18#01
BenchmarkJumpdestOpAnalysis/PUSH18#01-16     	    4789	    260621 ns/op	4714.90 MB/s
BenchmarkJumpdestOpAnalysis/PUSH19#01
BenchmarkJumpdestOpAnalysis/PUSH19#01-16     	    4610	    265413 ns/op	4629.77 MB/s
BenchmarkJumpdestOpAnalysis/PUSH20#01
BenchmarkJumpdestOpAnalysis/PUSH20#01-16     	    5389	    217422 ns/op	5651.68 MB/s
BenchmarkJumpdestOpAnalysis/PUSH21#01
BenchmarkJumpdestOpAnalysis/PUSH21#01-16     	    4698	    225808 ns/op	5441.80 MB/s
BenchmarkJumpdestOpAnalysis/PUSH22#01
BenchmarkJumpdestOpAnalysis/PUSH22#01-16     	    5725	    213327 ns/op	5760.18 MB/s
BenchmarkJumpdestOpAnalysis/PUSH23#01
BenchmarkJumpdestOpAnalysis/PUSH23#01-16     	    5410	    216495 ns/op	5675.89 MB/s
BenchmarkJumpdestOpAnalysis/PUSH24#01
BenchmarkJumpdestOpAnalysis/PUSH24#01-16     	    5870	    192940 ns/op	6368.83 MB/s
BenchmarkJumpdestOpAnalysis/PUSH25#01
BenchmarkJumpdestOpAnalysis/PUSH25#01-16     	    5488	    208259 ns/op	5900.34 MB/s
BenchmarkJumpdestOpAnalysis/PUSH26#01
BenchmarkJumpdestOpAnalysis/PUSH26#01-16     	    4882	    211021 ns/op	5823.11 MB/s
BenchmarkJumpdestOpAnalysis/PUSH27#01
BenchmarkJumpdestOpAnalysis/PUSH27#01-16     	    5646	    214649 ns/op	5724.70 MB/s
BenchmarkJumpdestOpAnalysis/PUSH28#01
BenchmarkJumpdestOpAnalysis/PUSH28#01-16     	    6424	    196888 ns/op	6241.10 MB/s
BenchmarkJumpdestOpAnalysis/PUSH29#01
BenchmarkJumpdestOpAnalysis/PUSH29#01-16     	    5580	    200527 ns/op	6127.86 MB/s
BenchmarkJumpdestOpAnalysis/PUSH30#01
BenchmarkJumpdestOpAnalysis/PUSH30#01-16     	    5847	    201044 ns/op	6112.08 MB/s
BenchmarkJumpdestOpAnalysis/PUSH31#01
BenchmarkJumpdestOpAnalysis/PUSH31#01-16     	    6160	    198827 ns/op	6180.25 MB/s
BenchmarkJumpdestOpAnalysis/PUSH32#01
BenchmarkJumpdestOpAnalysis/PUSH32#01-16     	    7820	    161729 ns/op	7597.90 MB/s
BenchmarkJumpdestOpAnalysis/JUMPDEST#01
BenchmarkJumpdestOpAnalysis/JUMPDEST#01-16   	    1940	    613130 ns/op	2004.14 MB/s
BenchmarkJumpdestOpAnalysis/STOP#01
BenchmarkJumpdestOpAnalysis/STOP#01-16       	    1873	    626232 ns/op	1962.21 MB/s
```